# ACCServer/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "shadowDemo restful API.",
    "version": "1.0"
  },
  "host": "localhost:8030",
  "basePath": "/acc",
  "paths": {
...
    "/iPWhiteList": {
      "get": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "iPWhiteList"
        ],
        "summary": "详情;",
        "operationId": "getIPWhiteList",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getIPWhiteListResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "put": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "iPWhiteList"
        ],
        "summary": "修改;",
        "operationId": "updateIPWhiteList",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "post": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "iPWhiteList"
        ],
        "summary": "创建;",
        "operationId": "createIPWhiteList",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "delete": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "iPWhiteList"
        ],
        "summary": "删除;",
        "operationId": "deleteIPWhiteList",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
    "/iPWhiteLists": {
      "post": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "iPWhiteList"
        ],
        "summary": "查询列表;",
        "operationId": "searchIPWhiteList",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "CreateEndTime": {
                  "description": "创建截止时间",
                  "type": "string",
                  "format": "date-time"
                },
                "CreateStartTime": {
                  "description": "创建开始时间",
                  "type": "string",
                  "format": "date-time"
                },
                "IDS": {
                  "description": "id",
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  }
                },
                "PageNum": {
                  "type": "integer",
                  "format": "int64"
                },
                "PageSize": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/searchIPWhiteListResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
...
    "/order": {
      "get": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "order"
        ],
        "summary": "详情;",
        "operationId": "getOrder",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getOrderResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "put": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "order"
        ],
        "summary": "修改;",
        "operationId": "updateOrder",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "post": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "order"
        ],
        "summary": "创建;",
        "operationId": "createOrder",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "delete": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "order"
        ],
        "summary": "删除;",
        "operationId": "deleteOrder",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
    "/orders": {
      "post": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "order"
        ],
        "summary": "查询列表;",
        "operationId": "searchOrder",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "CreateEndTime": {
                  "description": "创建截止时间",
                  "type": "string",
                  "format": "date-time"
                },
                "CreateStartTime": {
                  "description": "创建开始时间",
                  "type": "string",
                  "format": "date-time"
                },
                "IDS": {
                  "description": "id",
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  }
                },
                "PageNum": {
                  "type": "integer",
                  "format": "int64"
                },
                "PageSize": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/searchOrderResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Decimal": {
      "description": "number = value * 10 ^ exp",
      "type": "object",
      "title": "Decimal represents a fixed-point decimal. It is immutable.",
      "x-go-package": "github.com/shopspring/decimal"
    },
    "GenericMessageBody": {
      "type": "object",
      "properties": {
        "Code": {
          "type": "integer",
          "format": "int64"
        },
        "Msg": {
          "type": "string"
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
...
    "GetIPWhiteListResponseBody": {
      "type": "object",
      "properties": {
        "Result": {
          "$ref": "#/definitions/IPWhiteList"
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
    "GetOrderResponseBody": {
      "type": "object",
      "properties": {
        "Result": {
          "$ref": "#/definitions/Order"
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
...
    "IPWhiteList": {
      "type": "object",
      "properties": {
        "AccountName": {
          "description": "AccountType AccountType",
          "type": "string"
        },
        "AccountNo": {
          "type": "string"
        },
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "CreatedBy": {
          "type": "string"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "Enable": {
          "type": "boolean"
        },
        "ID": {
          "type": "integer",
          "format": "int64"
        },
        "IP": {
          "type": "string"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "shadowDemo/model/do"
    },
...
    "Order": {
      "type": "object",
      "properties": {
        "Abnormal": {
          "type": "boolean"
        },
        "AccountName": {
          "type": "string"
        },
        "AccountNo": {
          "type": "string"
        },
        "ApiVersion": {
          "type": "integer",
          "format": "int64"
        },
        "Bank": {
          "type": "string"
        },
        "BankCardCvv": {
          "type": "string"
        },
        "BankCardExpire": {
          "type": "string"
        },
        "BankCardNo": {
          "type": "string"
        },
        "CardOwner": {
          "type": "string"
        },
        "CardOwnerCert": {
          "description": "PayType             EnumPayType",
          "type": "string"
        },
        "CardOwnerPhone": {
          "type": "string"
        },
        "ChannelCode": {
          "type": "string"
        },
        "City": {
          "type": "string"
        },
        "ClientIP": {
          "type": "string"
        },
        "CodeMerchantOrder": {
          "type": "boolean"
        },
        "CreateTime": {
          "type": "string",
          "format": "date-time"
        },
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "CreatedBy": {
          "type": "string"
        },
        "CustomerNo": {
          "type": "string"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "int64"
        },
        "Information": {
          "type": "string"
        },
        "IsTest": {
          "description": "Payment             BccPayment `gorm:\"ForeignKey:PaymentID\"`",
          "type": "boolean"
        },
        "MerchantCallbackURL": {
          "type": "string"
        },
        "MerchantID": {
          "type": "integer",
          "format": "int64"
        },
        "MerchantName": {
          "type": "string"
        },
        "MerchantNo": {
          "type": "string"
        },
        "MerchantOrderNo": {
          "type": "string"
        },
        "MerchantSecretKey": {
          "type": "string"
        },
        "MerchantWithdrawFee": {
          "$ref": "#/definitions/Decimal"
        },
        "Operator": {
          "type": "string"
        },
        "OperatorNo": {
          "type": "string"
        },
        "OrderNo": {
          "type": "string"
        },
        "OrderPrice": {
          "$ref": "#/definitions/Decimal"
        },
        "OrderRealPrice": {
          "$ref": "#/definitions/Decimal"
        },
        "PaymentID": {
          "type": "integer",
          "format": "int64"
        },
        "Province": {
          "type": "string"
        },
        "RelatedOrderNo": {
          "description": "OrderType           OperateType     // 充值=1，提现=2\nOrderState          OrderState",
          "type": "string"
        },
        "Remark": {
          "type": "string"
        },
        "RemitType": {
          "type": "integer",
          "format": "int64"
        },
        "RiskLevel": {
          "type": "integer",
          "format": "int64"
        },
        "Subbranch": {
          "type": "string"
        },
        "UpdateTime": {
          "type": "string",
          "format": "date-time"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "shadowDemo/model/do"
    },
...
    "SearchIPWhiteListResponseBody": {
      "type": "object",
      "properties": {
        "Count": {
          "type": "integer",
          "format": "int64"
        },
        "Result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/IPWhiteList"
          }
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
    "SearchOrderResponseBody": {
      "type": "object",
      "properties": {
        "Count": {
          "type": "integer",
          "format": "int64"
        },
        "Result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Order"
          }
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
...
  },
  "responses": {
...
    "genericError": {
      "description": "错误消息",
      "schema": {
        "$ref": "#/definitions/GenericMessageBody"
      }
    },
    "genericSuccess": {
      "description": "成功消息",
      "schema": {
        "$ref": "#/definitions/GenericMessageBody"
      }
    },
...
    "getIPWhiteListResponse": {
      "description": "详情出参",
      "schema": {
        "$ref": "#/definitions/GetIPWhiteListResponseBody"
      }
    },
    "getOrderResponse": {
      "description": "详情出参",
      "schema": {
        "$ref": "#/definitions/GetOrderResponseBody"
      }
    },
...
    "searchIPWhiteListResponse": {
      "description": "查询列表",
      "schema": {
        "$ref": "#/definitions/SearchIPWhiteListResponseBody"
      }
    },
    "searchOrderResponse": {
      "description": "查询列表",
      "schema": {
        "$ref": "#/definitions/SearchOrderResponseBody"
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}